+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Reenable `Remote Write` to Cortex
 
+### Added
+
+- Trigger final heartbeat before deleting the cluster to clean up opened heartbeat alerts
+
 ## [1.15.0] - 2021-01-22
 
 ### Changed